% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrender.R
\name{xrender}
\alias{xrender}
\title{rmarkdown::render() for any spinscript}
\usage{
xrender(input, output = NULL, engine = NULL, keep_rmd = F, ...)
}
\arguments{
\item{input}{Input script name}

\item{output}{Path to compiled output.}

\item{engine}{Passed to xspin()}

\item{keep_rmd}{Should we keep the .Rmd file?}

\item{...}{Extra args passed to rmarkdown::render}
}
\description{
rmarkdown::render() is a user-friendly entry point to render an
Rmarkdown document. This function is an equivalent for a spin() script in
any language Rmarkdown supports.
}
