% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xspin.R
\name{xspin}
\alias{xspin}
\title{knitr::spin() for any language}
\usage{
xspin(input, output = fs::path_ext_set(input, ".Rmd"), engine = NULL)
}
\arguments{
\item{input}{Input script name}

\item{output}{Output Rmarkdown name (defaults to same basename as input
but with .Rmd extension)}

\item{engine}{Rmarkdown language engine to use. Defaults to guessing from
the extension of \code{input}.}
}
\description{
knitr::spin() translates a R script annotated with markdown
comments to a Rmarkdown notebook. At least currently, only R scripts are
supported by knitr::spin(), despite Rmarkdown having multi-language support.
This function is a much dumber version of spin() that doesn't do any
validation, but which supports any language supported by Rmarkdown code
cells.
}
